Shift out and dynamicise style, more logo updates
